opt: rename `exec/build` to `exec/execbuilder`
Release note: None
