+++ pkg/sql/opt/exec/execbuilder/builder_test.go
@@ -12,7 +12,7 @@
 // implied. See the License for the specific language governing
 // permissions and limitations under the License.
 
-package build
+package execbuilder
 
 import (
 	"bytes"
