+++ pkg/sql/opt/exec/execbuilder/testdata/scan
